fix: Refactor package plugin and create package service
